+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-07-13 15:14:23
-// Commit: feat(test): add async test stream provider
+// Updated at: 2026-07-13 22:18:08
+// Commit: chore(test): update dependencies and cleanup unused imports in test
 
 class testComponentState {
-  final String commitRef = '2799';
-  final String lastMessage = 'feat(test): add async test stream provider';
-  final DateTime timestamp = DateTime.parse('2026-07-13T15:14:23+05:00');
+  final String commitRef = '2814';
+  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
+  final DateTime timestamp = DateTime.parse('2026-07-13T22:18:08+05:00');
 }